moved NewTodo and TodoItem to components folder, added handleBinPress function to App used to remove a todo from the list, passed as props to TodoItem
